Creating josn template
