+++ convertToSingleHost.js
@@ -3,7 +3,7 @@
 const convertTest = process.argv[3] === "convert-test";
 const fs = require("fs");
 const host = process.argv[2];
-const hosts = ["excel", "onenote", "outlook", "powerpoint", "project", "word"];
+const hosts = ["outlook"];
 const path = require("path");
 const util = require("util");
 const testPackages = [
@@ -27,52 +27,16 @@ async function modifyProjectForSingleHost(host) {
   if (!hosts.includes(host)) {
     throw new Error(`'${host}' is not a supported host.`);
   }
-  await convertProjectToSingleHost(host);
+  await convertProjectToSingleHost();
   await updatePackageJsonForSingleHost(host);
   if (!convertTest) {
     await updateLaunchJsonFile();
   }
 }
 
-async function convertProjectToSingleHost(host) {
-  // copy host-specific manifest over manifest.xml
-  const manifestContent = await readFileAsync(`./manifest.${host}.xml`, "utf8");
-  await writeFileAsync(`./manifest.xml`, manifestContent);
-
-  // copy over host-specific taskpane code to taskpane.ts
-  const srcContent = await readFileAsync(`./src/taskpane/${host}.ts`, "utf8");
-  await writeFileAsync(`./src/taskpane/taskpane.ts`, srcContent);
-
+async function convertProjectToSingleHost() {
   // delete all test files by default for now - eventually we want to convert the tests by default
-  if (convertTest && (host === "excel" || host === "word")) {
-    // copy over host-specific taskpane test code to test-taskpane.ts
-    const testTaskpaneContent = await readFileAsync(`./test/src/${host}-test-taskpane.ts`, "utf8");
-    const updatedTestTaskpaneContent = testTaskpaneContent.replace(
-      `../../src/taskpane/${host}`,
-      `../../src/taskpane/taskpane`
-    );
-    await writeFileAsync(`./test/src/test-taskpane.ts`, updatedTestTaskpaneContent);
-
-    // update ui-test.ts to only run against specified host
-    const testContent = await readFileAsync(`./test/ui-test.ts`, "utf8");
-    const updatedTestContent = testContent.replace(`const hosts = ["Excel", "Word"]`, `const hosts = ["${host}"]`);
-    await writeFileAsync(`./test/ui-test.ts`, updatedTestContent);
-
-    // delete all host-specific test files after converting to single host
-    hosts.forEach(async function (host) {
-      if (host == "excel" || host == "word") {
-        await unlinkFileAsync(`./test/src/${host}-test-taskpane.ts`);
-      }
-    });
-  } else {
-    deleteFolder(path.resolve(`./test`));
-  }
-
-  // delete all host-specific files
-  hosts.forEach(async function (host) {
-    await unlinkFileAsync(`./manifest.${host}.xml`);
-    await unlinkFileAsync(`./src/taskpane/${host}.ts`);
-  });
+  deleteFolder(path.resolve(`./test`));
 
   // delete the .github folder
   deleteFolder(path.resolve(`./.github`));
