+++ package.json
@@ -23,15 +23,15 @@
     "lint": "office-addin-lint check",
     "lint:fix": "office-addin-lint fix",
     "prettier": "office-addin-lint prettier",
-    "start": "office-addin-debugging start manifest.xml",
-    "start:desktop": "office-addin-debugging start manifest.xml desktop",
-    "start:desktop:outlook": "office-addin-debugging start manifest.outlook.xml desktop --app outlook",
-    "start:web": "office-addin-debugging start manifest.xml web",
-    "stop": "office-addin-debugging stop manifest.xml",
+    "start": "office-addin-debugging start manifest.json",
+    "start:desktop": "office-addin-debugging start manifest.json desktop",
+    "start:desktop:outlook": "office-addin-debugging start manifest.outlook.json desktop --app outlook",
+    "start:web": "office-addin-debugging start manifest.json web",
+    "stop": "office-addin-debugging stop manifest.json",
     "test": "npm run test:unit && npm run test:e2e",
     "test:e2e": "mocha -r ts-node/register test/end-to-end/*.ts",
     "test:unit": "mocha -r ts-node/register test/unit/*.test.ts",
-    "validate": "office-addin-manifest validate manifest.xml",
+    "validate": "office-addin-manifest validate manifest.json",
     "watch": "webpack --mode development --watch"
   },
   "dependencies": {
