--- src/taskpane/onenote.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
- * See LICENSE in the project root for license information.
- */
-
-/* global document, Office */
-
-Office.onReady((info) => {
-  if (info.host === Office.HostType.OneNote) {
-    document.getElementById("sideload-msg").style.display = "none";
-    document.getElementById("app-body").style.display = "flex";
-    document.getElementById("run").onclick = run;
-  }
-});
-
-export async function run() {
-  /**
-   * Insert your OneNote code here
-   */
-}
