+++ src/taskpane/taskpane.ts
@@ -1,6 +1,20 @@
-import "./excel";
-import "./onenote";
-import "./outlook";
-import "./powerpoint";
-import "./project";
-import "./word";
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+ * See LICENSE in the project root for license information.
+ */
+
+/* global document, Office */
+
+Office.onReady((info) => {
+  if (info.host === Office.HostType.Outlook) {
+    document.getElementById("sideload-msg").style.display = "none";
+    document.getElementById("app-body").style.display = "flex";
+    document.getElementById("run").onclick = run;
+  }
+});
+
+export async function run() {
+  /**
+   * Insert your Outlook code here
+   */
+}
